+++ packages/nimble-components/src/card/tests/card.stories.ts
@@ -2,6 +2,7 @@ import { html } from '@microsoft/fast-element';
 import type { Meta, StoryObj } from '@storybook/html';
 import {
     createUserSelectedThemeStory,
+    disableStorybookZoomTransform,
     incubatingWarning
 } from '../../utilities/tests/storybook';
 import { listOptionTag } from '../../list-option';
@@ -9,10 +10,14 @@ import { numberFieldTag } from '../../number-field';
 import { selectTag } from '../../select';
 import { cardTag } from '..';
 
+interface CardArgs {
+    title: string;
+}
+
 const overviewText = `The \`nimble-card\` is a container that is designed to contain arbitrary content that is specified by a client
 application. The \`nimble-card\` is intended for grouping related content.`;
 
-const metadata: Meta = {
+const metadata: Meta<CardArgs> = {
     title: 'Incubating/Card',
     tags: ['autodocs'],
     parameters: {
@@ -24,11 +29,13 @@ const metadata: Meta = {
         actions: {}
     },
     render: createUserSelectedThemeStory(html`
+    ${disableStorybookZoomTransform}
     ${incubatingWarning({
         componentName: 'card',
         statusLink: 'https://github.com/ni/nimble/issues/296'
     })}
         <${cardTag}>
+            <span slot="title">${x => x.title}</span>
             <${numberFieldTag}>Numeric field 1</${numberFieldTag}>
             <${numberFieldTag}>Numeric field 2</${numberFieldTag}>
             <${selectTag}>
@@ -37,9 +44,18 @@ const metadata: Meta = {
                 <${listOptionTag} value="3">Option 3</${listOptionTag}>
             </${selectTag}>
         </${cardTag}>
-    `)
+    `),
+    argTypes: {
+        title: {
+            description:
+                'Text displayed as a title inside the card. Cards should **always include a title**. The title is used to provide an accessible name to assistive technologies.<br><br>Provide the title in an element targeted to the `title` slot.'
+        }
+    },
+    args: {
+        title: 'Title text'
+    }
 };
 
 export default metadata;
 
-export const card: StoryObj = {};
+export const card: StoryObj<CardArgs> = {};